provider-system/adapter/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==41.0.4
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.104.1
Flask==2.3.3
Flask-Cors==4.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==3.0.2
pika==1.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9.1
schedule==1.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
Werkzeug==3.1.3
zipp==3.21.0
python-keycloak
apscheduler